feat : implement user-service

+++ server/user-service/src/main/java/com/final_project_clinic/user/service/impl/PatientServiceImpl.java
@@ -64,7 +64,7 @@ public class PatientServiceImpl implements PatientService {
         existingPatient.setPhoneNumber(patientSaveDTO.getPhoneNumber());
         existingPatient.setDateOfBirth(patientSaveDTO.getDateOfBirth());
         existingPatient.setGender(patientSaveDTO.getGender());
-        existingPatient.setNik(patientSaveDTO.getNik());
+        existingPatient.setNik((long) patientSaveDTO.getNik());
         existingPatient.setUpdatedTime(new Date());
         existingPatient.setUpdatedBy("adm in"); // Replace with actual user from context
